Move some framework only commons into engine module part 2
