+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/eq/StringEq.kt
@@ -4,7 +4,7 @@ import io.kotest.assertions.*
 import io.kotest.assertions.print.Printed
 import io.kotest.assertions.print.print
 import io.kotest.assertions.print.printWithType
-import io.kotest.common.isIntellij
+import io.kotest.mpp.sysprop
 
 /**
  * An [Eq] implementation for String's that generates diffs for errors when the string inputs
@@ -30,8 +30,12 @@ object StringEq : Eq<String> {
                "(contents match, but line-breaks differ; output has been escaped to show line-breaks)\n"
             )
          }
+
          useDiff(expected, actual) -> diff(expected, actual)
-         else -> failureWithTypeInformation(ExpectedWithType(expected.printWithType()), ActualWithType(actual.printWithType()))
+         else -> failureWithTypeInformation(
+            ExpectedWithType(expected.printWithType()),
+            ActualWithType(actual.printWithType())
+         )
       }
    }
 
@@ -66,4 +70,15 @@ fun escapeLineBreaks(input: String): String {
       .replace("\r", "\\r")
 }
 
-
+/**
+ * Returns true if we are executing inside intellij.
+ *
+ * Note: This cannot be relied on for 100% accuracy.
+ */
+internal fun isIntellij(): Boolean {
+   return sysprop("idea.test.cyclic.buffer.size") != null
+      || (sysprop("jboss.modules.system.pkgs") ?: "").contains("com.intellij.rt")
+      || sysprop("intellij.debug.agent") != null
+      || (sysprop("java.class.path") ?: "").contains("idea_rt.jar")
+      || (sysprop("idea.active") != null)
+}
