+++ kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/factory/FactoryId.kt
@@ -1,7 +1,5 @@
 package io.kotest.core.factory
 
-import io.kotest.mpp.uniqueId
-
 data class FactoryId(val value: String) {
    companion object {
       fun next(): FactoryId = FactoryId(uniqueId())
