+++ kotest-framework/kotest-framework-api/src/jsHostedMain/kotlin/io/kotest/core/factory/uuid.kt
@@ -1,4 +1,4 @@
-package io.kotest.mpp
+package io.kotest.core.factory
 
 import kotlin.random.Random
